refactor(frontend): extract feed conversion pure modules
Split error normalization, preview hydration, and feeds API out of
useFeedConversion; reuse isNormalizedHttpUrl and nest preview status
on CreatedFeedResult for unit-testable policy seams.
